requirements.txt
pygame~=2.1.2
Pillow~=9.1.1